Work Session 2
Problem 3 is very interesting. I used the idea that the second smallest divisor of a number is always its prime factor to solve the problem. I first wanted to use recurrence but I discovered that the solution I came up with is even more concise. Problem 4 includes two situations and I wrote separated codes for each of the situation.