Work Session 3
For problem 5, I used the math way to solve the problem: the smallest common multiple of multiple numbers is the product of the greatest power of prime factors. After calculating the greatest power of prime factors, I used python as a calculator and found the answer. Problem 6 is about the use of for loop. After using range(1,100) and finding the sum of squares and the sum, we square the sum and end up with the sum of squares and the square of the sum. Finally, we use absolute value function to find their difference.