Work Session 4
Problem 7 is purely about finding prime numbers in the fastest way. What I did is to put all prime numbers that I've found so far into a list and moving forward, I compare a number with what I currently have in the list and determine whether it is a prime or not. Using this method, we can easily find out the 10001st prime number. Ms. DeRanek also suggested a way to improve my code, which is to only use the prime numbers smaller than the square root of a number to determine whether it is prime or not. Because the smallest prime factor of a number cannot be greater than its square root.