Work Session 5
For problem 8, I located every 0 in the number and only calculated the product of adjacent numbers in between 0s, that way I can reduce the runtime as any operation of multiplication that involves 0 ends up with 0. Then I compared every product of 13 adjacent numbers and came up with the greatest product as my answer.