Work Session 6
For problem 9, I used double for loop to come up with every possible triplets of (a, b, c) and find the only triplets that fulfills a^2+b^2=c^2. For problem 10, I used a similar method as problem 7 to find every prime number under 2000001 and added them together. The only thing that I've improved on the method in problem 7 is that I used the property that the least prime factor of a number is smaller than its square root to reduce runtime.